Camera movement improved
Added CustomAvatars with an avatar that transforms the lookup parameters (position, target and up) to position and direction.
Added a RelativeUniformLinearMotion behavior that transforms the input velocity in another that depends on the direction the objet is looking to.
Some changes to use the previous changes to move the camera.